pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "news-collector"
version = "0.1.0"
description = "고품질 뉴스 수집 시스템"
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["news_collector*"]
